AIAgent_GH/AI_Agent/requirements.txt
langchain 
langchain-community 
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
pandas
matplotlib
tempfile